refactor: Extract useFilesWithPath hook
