+++ src/drive/web/modules/views/Recent/index.jsx
@@ -1,8 +1,4 @@
 import React, { useCallback, useContext } from 'react'
-import { useQuery } from 'cozy-client'
-import get from 'lodash/get'
-import uniq from 'lodash/uniq'
-import keyBy from 'lodash/keyBy'
 
 import SharingProvider from 'cozy-sharing'
 import {
@@ -30,42 +26,16 @@ import Breadcrumb from './Breadcrumb'
 import File from '../FileWithActions'
 import RealTimeQueries from '../Drive/RealTimeQueries'
 
-import {
-  buildRecentQuery,
-  buildParentsByIdsQuery
-} from 'drive/web/modules/queries'
+import { buildRecentQuery } from 'drive/web/modules/queries'
+import { useFilesQueryWithPath } from './useFilesQueryWithPath'
 
-const useFilesQueryWithPath = query => {
-  const result = useQuery(query.definition, query.options)
-  const resultData = result.data || []
-
-  const dirIds = uniq(resultData.map(({ dir_id }) => dir_id))
-  const parentsQuery = buildParentsByIdsQuery(dirIds)
-  const parentsResult = useQuery(parentsQuery.definition, parentsQuery.options)
-
-  let parentsDocsById = {}
-
-  if (parentsResult.fetchStatus === 'loaded' && parentsResult.data.length > 0) {
-    parentsDocsById = keyBy(parentsResult.data, '_id')
-  }
-
-  return {
-    ...result,
-    data: resultData.map(file => ({
-      ...file,
-      displayedPath: get(parentsDocsById[file.dir_id], 'path')
-    }))
-  }
-}
-
-const RecentView = ({ router, children }) => {
+export const RecentView = ({ router, children }) => {
   const { isBigThumbnail, toggleThumbnailSize } = useContext(
     ThumbnailSizeContext
   )
 
   const query = buildRecentQuery()
   const result = useFilesQueryWithPath(query)
-
   const isInError = result.fetchStatus === 'error'
   const hasDataToShow = !isInError && (result.data && result.data.length > 0)
   const isLoading = !hasDataToShow && result.fetchStatus === 'loading'
